week3/weather-mcp-server/pyproject.toml
[project]
name = "weather-mcp-server"
version = "0.1.0"
description = "Weather MCP Server using OpenWeather API"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=0.2.0",
    "httpx>=0.27.0",
    "uvicorn>=0.30.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
]

